+++ examples/data-sources/nxos_pim_ssm_range/data-source.tf
@@ -1,0 +1,3 @@
+data "nxos_pim_ssm_range" "example" {
+  vrf_name = "default"
+}