+++ examples/resources/nxos_pim_ssm_policy/import.sh
@@ -1,0 +1,1 @@
+terraform import nxos_pim_ssm_policy.example "sys/pim/inst/dom-[default]/ssm"