+++ examples/resources/nxos_pim_ssm_policy/resource.tf
@@ -1,0 +1,4 @@
+resource "nxos_pim_ssm_policy" "example" {
+  vrf_name = "default"
+  name     = "SSM"
+}