+++ examples/resources/nxos_pim_ssm_range/import.sh
@@ -1,0 +1,1 @@
+terraform import nxos_pim_ssm_range.example "sys/pim/inst/dom-[default]/ssm/range"