+++ examples/resources/nxos_pim_static_rp_group_list/import.sh
@@ -1,0 +1,1 @@
+terraform import nxos_pim_static_rp_group_list.example "sys/pim/inst/dom-[default]/staticrp/rp-[1.2.3.4]/rpgrplist-[224.0.0.0/4]"